Adding reusable component
