+++ components/blocks/index.ts
@@ -32,6 +32,7 @@ import { flagSchema } from "./flag";
 import { googleMapsSchema } from "./googleMapsWrapper";
 import { heroBlockSchema } from "./hero";
 import { internalCarouselBlockSchema } from "./internalCarousel";
+import { jotFormEmbedSchema } from "./jotFormEmbed";
 import { newslettersTableBlockSchema } from "./newslettersTable";
 import { paymentBlockSchema } from "./payment-block";
 import { recurringEventSchema } from "./recurringEvent";
@@ -86,6 +87,7 @@ export const pageBlocks: Template[] = [
   latestTechSchema,
   testimonialsListSchema,
   eventLinkSchema,
+  jotFormEmbedSchema,
 ];
 
 export * from "../bookingButton/bookingButton";
@@ -115,6 +117,7 @@ export * from "./fixedColumns";
 export * from "./flag";
 export * from "./googleMapsWrapper";
 export * from "./hero";
+export * from "./jotFormEmbed";
 export * from "./newslettersTable";
 export * from "./recurringEvent";
 export * from "./serviceCards";
